+++ doc/CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 - Added optional command line arg to specify length of unique molecular tag (UMT)
 - Added optional command line arg to sort filters results by name instead of count
 - Added validation to check for properly paired alignments 
+- Added validation to check for presence of secondary alignments 
 - Adjusted so warning instead of error when no families found
 - Substantial refactors to clarify implementation